nda-validator-project/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==1.0.1
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.3
python-jose==3.3.0
passlib==1.7.4
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
langchain==0.0.350
chromadb==0.4.18 